requirements.txt
torch~=2.7.1
Flask~=3.1.1
torchvision~=0.22.1
pillow~=11.3.0